+++ src/components/ChatMessage.jsx
@@ -8,7 +8,7 @@ import { t } from '../i18n';
 import { isPlanApprovalPrompt } from './ChatView';
 import DiffView from './DiffView';
 import ToolResultView from './ToolResultView';
-import TranslateTag from './TranslateTag';
+
 import defaultAvatarUrl from '../img/default-avatar.svg';
 import defaultModelAvatarUrl from '../img/default-model-avatar.svg';
 import styles from './ChatMessage.module.css';
@@ -28,8 +28,6 @@ class ChatMessage extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      // { [thinkingIndex]: translatedHtml | null }
-      thinkingTranslated: {},
       planFeedbackInput: false,
       planFeedbackText: '',
       planFeedbackOptNumber: null,
@@ -47,15 +45,6 @@ class ChatMessage extends React.Component {
       });
     }
   }
-
-  handleThinkingTranslated = (index, translatedText) => {
-    this.setState(prev => ({
-      thinkingTranslated: {
-        ...prev.thinkingTranslated,
-        [index]: translatedText ? renderMarkdown(translatedText) : null,
-      },
-    }));
-  };
 
   formatTime(ts) {
     if (!ts) return null;
@@ -595,8 +584,6 @@ class ChatMessage extends React.Component {
     let innerContent = [];
 
     thinkingBlocks.forEach((tb, i) => {
-      const translatedHtml = this.state.thinkingTranslated[i];
-      const displayHtml = translatedHtml || renderMarkdown(tb.thinking || '');
       innerContent.push(
         <Collapse
           key={`think-${i}-${this.props.expandThinking ? 'e' : 'c'}`}
@@ -605,13 +592,8 @@ class ChatMessage extends React.Component {
           defaultActiveKey={this.props.expandThinking ? ['1'] : []}
           items={[{
             key: '1',
-            label: (
-              <span>
-                <Text type="secondary" className={styles.thinkingLabel}>{t('ui.thinking')}</Text>
-                <TranslateTag text={tb.thinking || ''} onTranslated={(txt) => this.handleThinkingTranslated(i, txt)} />
-              </span>
-            ),
-            children: <div className="chat-md" dangerouslySetInnerHTML={{ __html: displayHtml }} />,
+            label: <Text type="secondary" className={styles.thinkingLabel}>{t('ui.thinking')}</Text>,
+            children: <div className="chat-md" dangerouslySetInnerHTML={{ __html: renderMarkdown(tb.thinking || '') }} />,
           }]}
           className={styles.collapseMargin}
         />